fix(65-02): apply clippy auto-fixes for code quality
- Reduced warnings from 194 to 127 (67 auto-fixed)
- Fixed unused imports (PathBuf, Rc, HashMap, etc.)
- Fixed unused variables (prefixed with underscore)
- Fixed redundant field names in struct initialization
- Fixed borrowed expression traits warnings
- Fixed unnecessary .clone() calls
- Fixed io_other_error with std::io::Error::other()
- Fixed if_same_then_else in verify_cmd.rs

+++ src/graph/references.rs
@@ -141,12 +141,12 @@ impl ReferenceOps {
                                     fqn: if fqn.is_empty() { None } else { Some(fqn) },
                                     canonical_fqn: None,
                                     display_fqn: None,
-                                    byte_start: symbol_node.byte_start as usize,
-                                    byte_end: symbol_node.byte_end as usize,
-                                    start_line: symbol_node.start_line as usize,
-                                    start_col: symbol_node.start_col as usize,
-                                    end_line: symbol_node.end_line as usize,
-                                    end_col: symbol_node.end_col as usize,
+                                    byte_start: symbol_node.byte_start,
+                                    byte_end: symbol_node.byte_end,
+                                    start_line: symbol_node.start_line,
+                                    start_col: symbol_node.start_col,
+                                    end_line: symbol_node.end_line,
+                                    end_col: symbol_node.end_col,
                                 });
                             }
                         }
@@ -276,12 +276,12 @@ impl ReferenceOps {
                                     fqn: if fqn.is_empty() { None } else { Some(fqn) },
                                     canonical_fqn: symbol_node.canonical_fqn.clone(),
                                     display_fqn: symbol_node.display_fqn.clone(),
-                                    byte_start: symbol_node.byte_start as usize,
-                                    byte_end: symbol_node.byte_end as usize,
-                                    start_line: symbol_node.start_line as usize,
-                                    start_col: symbol_node.start_col as usize,
-                                    end_line: symbol_node.end_line as usize,
-                                    end_col: symbol_node.end_col as usize,
+                                    byte_start: symbol_node.byte_start,
+                                    byte_end: symbol_node.byte_end,
+                                    start_line: symbol_node.start_line,
+                                    start_col: symbol_node.start_col,
+                                    end_line: symbol_node.end_line,
+                                    end_col: symbol_node.end_col,
                                 });
                             }
                         }
