+++ src/ingest/cpp.rs
@@ -80,7 +80,7 @@ impl CppParser {
             let mut cursor = node.walk();
             for child in node.children(&mut cursor) {
                 if child.kind() == "namespace_identifier" {
-                    let name_bytes = safe_slice(source, child.start_byte() as usize, child.end_byte() as usize)?;
+                    let name_bytes = safe_slice(source, child.start_byte(), child.end_byte())?;
                     return std::str::from_utf8(name_bytes).ok().map(|s| s.to_string());
                 }
             }
@@ -97,7 +97,7 @@ impl CppParser {
         for child in node.children(&mut cursor) {
             match child.kind() {
                 "identifier" | "type_identifier" => {
-                    let name_bytes = safe_slice(source, child.start_byte() as usize, child.end_byte() as usize)?;
+                    let name_bytes = safe_slice(source, child.start_byte(), child.end_byte())?;
                     return std::str::from_utf8(name_bytes).ok().map(|s| s.to_string());
                 }
                 // Skip certain nodes to find the identifier within
@@ -274,8 +274,8 @@ impl CppParser {
             fqn: Some(fqn),
             canonical_fqn: Some(canonical_fqn),
             display_fqn: Some(display_fqn),
-            byte_start: node.start_byte() as usize,
-            byte_end: node.end_byte() as usize,
+            byte_start: node.start_byte(),
+            byte_end: node.end_byte(),
             start_line: node.start_position().row + 1,
             start_col: node.start_position().column,
             end_line: node.end_position().row + 1,
@@ -294,7 +294,7 @@ impl CppParser {
             let mut cursor = node.walk();
             for child in node.children(&mut cursor) {
                 if child.kind() == "namespace_identifier" {
-                    let name_bytes = safe_slice(source, child.start_byte() as usize, child.end_byte() as usize)?;
+                    let name_bytes = safe_slice(source, child.start_byte(), child.end_byte())?;
                     return std::str::from_utf8(name_bytes).ok().map(|s| s.to_string());
                 }
             }
@@ -312,7 +312,7 @@ impl CppParser {
             match child.kind() {
                 "identifier" | "type_identifier" => {
                     let name_bytes =
-                        &source[child.start_byte() as usize..child.end_byte() as usize];
+                        &source[child.start_byte()..child.end_byte()];
                     return std::str::from_utf8(name_bytes).ok().map(|s| s.to_string());
                 }
                 // Skip certain nodes to find the identifier within
@@ -450,8 +450,8 @@ impl CppParser {
             fqn: Some(fqn),
             canonical_fqn: Some(canonical_fqn),
             display_fqn: Some(display_fqn),
-            byte_start: node.start_byte() as usize,
-            byte_end: node.end_byte() as usize,
+            byte_start: node.start_byte(),
+            byte_end: node.end_byte(),
             start_line: node.start_position().row + 1,
             start_col: node.start_position().column,
             end_line: node.end_position().row + 1,
@@ -505,14 +505,14 @@ impl CppParser {
             return None;
         }
 
-        let text_bytes = &source[node.start_byte() as usize..node.end_byte() as usize];
+        let text_bytes = &source[node.start_byte()..node.end_byte()];
         let text = std::str::from_utf8(text_bytes).ok()?;
 
         let referenced_symbol = symbols
             .iter()
             .find(|s| s.name.as_ref().map(|n| n == text).unwrap_or(false))?;
 
-        let ref_start = node.start_byte() as usize;
+        let ref_start = node.start_byte();
         if ref_start < referenced_symbol.byte_end {
             return None;
         }
@@ -521,7 +521,7 @@ impl CppParser {
             file_path: file_path.clone(),
             referenced_symbol: text.to_string(),
             byte_start: ref_start,
-            byte_end: node.end_byte() as usize,
+            byte_end: node.end_byte(),
             start_line: node.start_position().row + 1,
             start_col: node.start_position().column,
             end_line: node.end_position().row + 1,
@@ -625,8 +625,8 @@ impl CppParser {
         if node.kind() == "call_expression" {
             if let Some(callee_name) = self.extract_callee_from_call(node, source) {
                 if symbol_map.contains_key(&callee_name) {
-                    let node_start = node.start_byte() as usize;
-                    let node_end = node.end_byte() as usize;
+                    let node_start = node.start_byte();
+                    let node_end = node.end_byte();
                     let call_fact = CallFact {
                         file_path: file_path.clone(),
                         caller: caller.name.clone().unwrap_or_default(),
@@ -650,7 +650,7 @@ impl CppParser {
         let mut cursor = node.walk();
         for child in node.children(&mut cursor) {
             if child.kind() == "identifier" {
-                let name_bytes = &source[child.start_byte() as usize..child.end_byte() as usize];
+                let name_bytes = &source[child.start_byte()..child.end_byte()];
                 return std::str::from_utf8(name_bytes).ok().map(|s| s.to_string());
             }
         }
