+++ src/ingest/java.rs
@@ -203,8 +203,8 @@ impl JavaParser {
             fqn: Some(fqn),
             canonical_fqn: Some(canonical_fqn),
             display_fqn: Some(display_fqn),
-            byte_start: node.start_byte() as usize,
-            byte_end: node.end_byte() as usize,
+            byte_start: node.start_byte(),
+            byte_end: node.end_byte(),
             start_line: node.start_position().row + 1, // tree-sitter is 0-indexed
             start_col: node.start_position().column,
             end_line: node.end_position().row + 1,
@@ -229,7 +229,7 @@ impl JavaParser {
             let mut cursor = node.walk();
             for child in node.children(&mut cursor) {
                 if child.kind() == "scoped_identifier" || child.kind() == "identifier" {
-                    let name_bytes = safe_slice(source, child.start_byte() as usize, child.end_byte() as usize)?;
+                    let name_bytes = safe_slice(source, child.start_byte(), child.end_byte())?;
                     return std::str::from_utf8(name_bytes).ok().map(|s| s.to_string());
                 }
             }
@@ -240,7 +240,7 @@ impl JavaParser {
         let mut cursor = node.walk();
         for child in node.children(&mut cursor) {
             if child.kind() == "identifier" {
-                let name_bytes = safe_slice(source, child.start_byte() as usize, child.end_byte() as usize)?;
+                let name_bytes = safe_slice(source, child.start_byte(), child.end_byte())?;
                 return std::str::from_utf8(name_bytes).ok().map(|s| s.to_string());
             }
         }
@@ -422,8 +422,8 @@ impl JavaParser {
             fqn: Some(fqn),
             canonical_fqn: Some(canonical_fqn),
             display_fqn: Some(display_fqn),
-            byte_start: node.start_byte() as usize,
-            byte_end: node.end_byte() as usize,
+            byte_start: node.start_byte(),
+            byte_end: node.end_byte(),
             start_line: node.start_position().row + 1,
             start_col: node.start_position().column,
             end_line: node.end_position().row + 1,
@@ -442,7 +442,7 @@ impl JavaParser {
             let mut cursor = node.walk();
             for child in node.children(&mut cursor) {
                 if child.kind() == "scoped_identifier" || child.kind() == "identifier" {
-                    let name_bytes = safe_slice(source, child.start_byte() as usize, child.end_byte() as usize)?;
+                    let name_bytes = safe_slice(source, child.start_byte(), child.end_byte())?;
                     return std::str::from_utf8(name_bytes).ok().map(|s| s.to_string());
                 }
             }
@@ -453,7 +453,7 @@ impl JavaParser {
         let mut cursor = node.walk();
         for child in node.children(&mut cursor) {
             if child.kind() == "identifier" {
-                let name_bytes = safe_slice(source, child.start_byte() as usize, child.end_byte() as usize)?;
+                let name_bytes = safe_slice(source, child.start_byte(), child.end_byte())?;
                 return std::str::from_utf8(name_bytes).ok().map(|s| s.to_string());
             }
         }
@@ -507,14 +507,14 @@ impl JavaParser {
             return None;
         }
 
-        let text_bytes = &source[node.start_byte() as usize..node.end_byte() as usize];
+        let text_bytes = &source[node.start_byte()..node.end_byte()];
         let text = std::str::from_utf8(text_bytes).ok()?;
 
         let referenced_symbol = symbols
             .iter()
             .find(|s| s.name.as_ref().map(|n| n == text).unwrap_or(false))?;
 
-        let ref_start = node.start_byte() as usize;
+        let ref_start = node.start_byte();
         if ref_start < referenced_symbol.byte_end {
             return None;
         }
@@ -523,7 +523,7 @@ impl JavaParser {
             file_path: file_path.clone(),
             referenced_symbol: text.to_string(),
             byte_start: ref_start,
-            byte_end: node.end_byte() as usize,
+            byte_end: node.end_byte(),
             start_line: node.start_position().row + 1,
             start_col: node.start_position().column,
             end_line: node.end_position().row + 1,
@@ -615,7 +615,7 @@ impl JavaParser {
         let mut cursor = node.walk();
         for child in node.children(&mut cursor) {
             if child.kind() == "identifier" {
-                let name_bytes = safe_slice(source, child.start_byte() as usize, child.end_byte() as usize)?;
+                let name_bytes = safe_slice(source, child.start_byte(), child.end_byte())?;
                 return std::str::from_utf8(name_bytes).ok().map(|s| s.to_string());
             }
         }
@@ -634,8 +634,8 @@ impl JavaParser {
         if node.kind() == "method_invocation" {
             if let Some(callee_name) = self.extract_callee_from_call(node, source) {
                 if symbol_map.contains_key(&callee_name) {
-                    let node_start = node.start_byte() as usize;
-                    let node_end = node.end_byte() as usize;
+                    let node_start = node.start_byte();
+                    let node_end = node.end_byte();
                     let call_fact = CallFact {
                         file_path: file_path.clone(),
                         caller: caller.name.clone().unwrap_or_default(),
@@ -659,7 +659,7 @@ impl JavaParser {
         let mut cursor = node.walk();
         for child in node.children(&mut cursor) {
             if child.kind() == "identifier" {
-                let name_bytes = safe_slice(source, child.start_byte() as usize, child.end_byte() as usize)?;
+                let name_bytes = safe_slice(source, child.start_byte(), child.end_byte())?;
                 return std::str::from_utf8(name_bytes).ok().map(|s| s.to_string());
             }
         }
