+++ src/ingest/javascript.rs
@@ -181,8 +181,8 @@ impl JavaScriptParser {
             fqn: Some(fqn),
             canonical_fqn: Some(canonical_fqn),
             display_fqn: Some(display_fqn),
-            byte_start: node.start_byte() as usize,
-            byte_end: node.end_byte() as usize,
+            byte_start: node.start_byte(),
+            byte_end: node.end_byte(),
             start_line: node.start_position().row + 1, // tree-sitter is 0-indexed
             start_col: node.start_position().column,
             end_line: node.end_position().row + 1,
@@ -198,7 +198,7 @@ impl JavaScriptParser {
         for child in node.children(&mut cursor) {
             match child.kind() {
                 "identifier" | "property_identifier" => {
-                    let name_bytes = safe_slice(source, child.start_byte() as usize, child.end_byte() as usize)?;
+                    let name_bytes = safe_slice(source, child.start_byte(), child.end_byte())?;
                     return std::str::from_utf8(name_bytes).ok().map(|s| s.to_string());
                 }
                 _ => {}
@@ -361,8 +361,8 @@ impl JavaScriptParser {
             fqn: Some(fqn),
             canonical_fqn: Some(canonical_fqn),
             display_fqn: Some(display_fqn),
-            byte_start: node.start_byte() as usize,
-            byte_end: node.end_byte() as usize,
+            byte_start: node.start_byte(),
+            byte_end: node.end_byte(),
             start_line: node.start_position().row + 1,
             start_col: node.start_position().column,
             end_line: node.end_position().row + 1,
@@ -378,7 +378,7 @@ impl JavaScriptParser {
         for child in node.children(&mut cursor) {
             match child.kind() {
                 "identifier" | "property_identifier" => {
-                    let name_bytes = safe_slice(source, child.start_byte() as usize, child.end_byte() as usize)?;
+                    let name_bytes = safe_slice(source, child.start_byte(), child.end_byte())?;
                     return std::str::from_utf8(name_bytes).ok().map(|s| s.to_string());
                 }
                 _ => {}
@@ -452,7 +452,7 @@ impl JavaScriptParser {
         }
 
         // Get the text of this node
-        let text_bytes = &source[node.start_byte() as usize..node.end_byte() as usize];
+        let text_bytes = &source[node.start_byte()..node.end_byte()];
         let text = std::str::from_utf8(text_bytes).ok()?;
 
         // Find if this matches any symbol
@@ -461,7 +461,7 @@ impl JavaScriptParser {
             .find(|s| s.name.as_ref().map(|n| n == text).unwrap_or(false))?;
 
         // Check if reference is OUTSIDE the symbol's defining span
-        let ref_start = node.start_byte() as usize;
+        let ref_start = node.start_byte();
 
         // Reference must start after the symbol's definition ends
         if ref_start < referenced_symbol.byte_end {
@@ -472,7 +472,7 @@ impl JavaScriptParser {
             file_path: file_path.clone(),
             referenced_symbol: text.to_string(),
             byte_start: ref_start,
-            byte_end: node.end_byte() as usize,
+            byte_end: node.end_byte(),
             start_line: node.start_position().row + 1,
             start_col: node.start_position().column,
             end_line: node.end_position().row + 1,
@@ -583,7 +583,7 @@ impl JavaScriptParser {
         let mut cursor = node.walk();
         for child in node.children(&mut cursor) {
             if child.kind() == "identifier" {
-                let name_bytes = safe_slice(source, child.start_byte() as usize, child.end_byte() as usize)?;
+                let name_bytes = safe_slice(source, child.start_byte(), child.end_byte())?;
                 return std::str::from_utf8(name_bytes).ok().map(|s| s.to_string());
             }
         }
@@ -608,8 +608,8 @@ impl JavaScriptParser {
             if let Some(callee_name) = self.extract_callee_from_call(node, source) {
                 // Only create call if callee is a known function symbol
                 if symbol_map.contains_key(&callee_name) {
-                    let node_start = node.start_byte() as usize;
-                    let node_end = node.end_byte() as usize;
+                    let node_start = node.start_byte();
+                    let node_end = node.end_byte();
                     let call_fact = CallFact {
                         file_path: file_path.clone(),
                         caller: caller.name.clone().unwrap_or_default(),
@@ -635,7 +635,7 @@ impl JavaScriptParser {
         let mut cursor = node.walk();
         for child in node.children(&mut cursor) {
             if child.kind() == "identifier" {
-                let name_bytes = safe_slice(source, child.start_byte() as usize, child.end_byte() as usize)?;
+                let name_bytes = safe_slice(source, child.start_byte(), child.end_byte())?;
                 return std::str::from_utf8(name_bytes).ok().map(|s| s.to_string());
             }
             // Handle member_expression calls like obj.method() - we want the method name
@@ -659,7 +659,7 @@ impl JavaScriptParser {
             // Find property_identifier (second child in obj.prop)
             let prop = &children[1];
             if prop.kind() == "property_identifier" {
-                let name_bytes = safe_slice(source, prop.start_byte() as usize, prop.end_byte() as usize)?;
+                let name_bytes = safe_slice(source, prop.start_byte(), prop.end_byte())?;
                 return std::str::from_utf8(name_bytes).ok().map(|s| s.to_string());
             }
         }
