+++ src/ingest/typescript.rs
@@ -191,8 +191,8 @@ impl TypeScriptParser {
             fqn: Some(fqn),
             canonical_fqn: Some(canonical_fqn),
             display_fqn: Some(display_fqn),
-            byte_start: node.start_byte() as usize,
-            byte_end: node.end_byte() as usize,
+            byte_start: node.start_byte(),
+            byte_end: node.end_byte(),
             start_line: node.start_position().row + 1, // tree-sitter is 0-indexed
             start_col: node.start_position().column,
             end_line: node.end_position().row + 1,
@@ -212,7 +212,7 @@ impl TypeScriptParser {
             let mut cursor = node.walk();
             for child in node.children(&mut cursor) {
                 if child.kind() == "identifier" {
-                    let name_bytes = safe_slice(source, child.start_byte() as usize, child.end_byte() as usize)?;
+                    let name_bytes = safe_slice(source, child.start_byte(), child.end_byte())?;
                     return std::str::from_utf8(name_bytes).ok().map(|s| s.to_string());
                 }
             }
@@ -224,7 +224,7 @@ impl TypeScriptParser {
         for child in node.children(&mut cursor) {
             match child.kind() {
                 "identifier" | "type_identifier" | "property_identifier" => {
-                    let name_bytes = safe_slice(source, child.start_byte() as usize, child.end_byte() as usize)?;
+                    let name_bytes = safe_slice(source, child.start_byte(), child.end_byte())?;
                     return std::str::from_utf8(name_bytes).ok().map(|s| s.to_string());
                 }
                 _ => {}
@@ -396,8 +396,8 @@ impl TypeScriptParser {
             fqn: Some(fqn),
             canonical_fqn: Some(canonical_fqn),
             display_fqn: Some(display_fqn),
-            byte_start: node.start_byte() as usize,
-            byte_end: node.end_byte() as usize,
+            byte_start: node.start_byte(),
+            byte_end: node.end_byte(),
             start_line: node.start_position().row + 1,
             start_col: node.start_position().column,
             end_line: node.end_position().row + 1,
@@ -416,7 +416,7 @@ impl TypeScriptParser {
             let mut cursor = node.walk();
             for child in node.children(&mut cursor) {
                 if child.kind() == "identifier" {
-                    let name_bytes = safe_slice(source, child.start_byte() as usize, child.end_byte() as usize)?;
+                    let name_bytes = safe_slice(source, child.start_byte(), child.end_byte())?;
                     return std::str::from_utf8(name_bytes).ok().map(|s| s.to_string());
                 }
             }
@@ -428,7 +428,7 @@ impl TypeScriptParser {
         for child in node.children(&mut cursor) {
             match child.kind() {
                 "identifier" | "type_identifier" | "property_identifier" => {
-                    let name_bytes = safe_slice(source, child.start_byte() as usize, child.end_byte() as usize)?;
+                    let name_bytes = safe_slice(source, child.start_byte(), child.end_byte())?;
                     return std::str::from_utf8(name_bytes).ok().map(|s| s.to_string());
                 }
                 _ => {}
@@ -502,7 +502,7 @@ impl TypeScriptParser {
         }
 
         // Get the text of this node
-        let text_bytes = safe_slice(source, node.start_byte() as usize, node.end_byte() as usize)?;
+        let text_bytes = safe_slice(source, node.start_byte(), node.end_byte())?;
         let text = std::str::from_utf8(text_bytes).ok()?;
 
         // Find if this matches any symbol
@@ -511,7 +511,7 @@ impl TypeScriptParser {
             .find(|s| s.name.as_ref().map(|n| n == text).unwrap_or(false))?;
 
         // Check if reference is OUTSIDE the symbol's defining span
-        let ref_start = node.start_byte() as usize;
+        let ref_start = node.start_byte();
 
         // Reference must start after the symbol's definition ends
         if ref_start < referenced_symbol.byte_end {
@@ -522,7 +522,7 @@ impl TypeScriptParser {
             file_path: file_path.clone(),
             referenced_symbol: text.to_string(),
             byte_start: ref_start,
-            byte_end: node.end_byte() as usize,
+            byte_end: node.end_byte(),
             start_line: node.start_position().row + 1,
             start_col: node.start_position().column,
             end_line: node.end_position().row + 1,
@@ -632,7 +632,7 @@ impl TypeScriptParser {
         let mut cursor = node.walk();
         for child in node.children(&mut cursor) {
             if child.kind() == "identifier" {
-                let name_bytes = safe_slice(source, child.start_byte() as usize, child.end_byte() as usize)?;
+                let name_bytes = safe_slice(source, child.start_byte(), child.end_byte())?;
                 return std::str::from_utf8(name_bytes).ok().map(|s| s.to_string());
             }
         }
@@ -657,8 +657,8 @@ impl TypeScriptParser {
             if let Some(callee_name) = self.extract_callee_from_call(node, source) {
                 // Only create call if callee is a known function symbol
                 if symbol_map.contains_key(&callee_name) {
-                    let node_start = node.start_byte() as usize;
-                    let node_end = node.end_byte() as usize;
+                    let node_start = node.start_byte();
+                    let node_end = node.end_byte();
                     let call_fact = CallFact {
                         file_path: file_path.clone(),
                         caller: caller.name.clone().unwrap_or_default(),
@@ -684,7 +684,7 @@ impl TypeScriptParser {
         let mut cursor = node.walk();
         for child in node.children(&mut cursor) {
             if child.kind() == "identifier" {
-                let name_bytes = safe_slice(source, child.start_byte() as usize, child.end_byte() as usize)?;
+                let name_bytes = safe_slice(source, child.start_byte(), child.end_byte())?;
                 return std::str::from_utf8(name_bytes).ok().map(|s| s.to_string());
             }
             // Handle member_expression calls like obj.method() - we want the property name
@@ -704,7 +704,7 @@ impl TypeScriptParser {
             // For member_expression, look for property_identifier
             for child in &children {
                 if child.kind() == "property_identifier" {
-                    let name_bytes = safe_slice(source, child.start_byte() as usize, child.end_byte() as usize)?;
+                    let name_bytes = safe_slice(source, child.start_byte(), child.end_byte())?;
                     return std::str::from_utf8(name_bytes).ok().map(|s| s.to_string());
                 }
             }
